Increment given data pointer on primitive parsing
This was supposed to be done before but I forgor since I hadn't tested it

+++ libs/cubic-protocol/include/cubic-protocol/primitives/string.hpp
@@ -12,9 +12,9 @@ namespace cubic::protocol::primitives::string {
 auto parse(const uint8_t *data, uint32_t available_bytes, std::string *value) -> uint32_t;
 auto parse(uint8_t *const *data, uint32_t available_bytes) -> std::optional<std::string>;
 auto parse(
-    const uint8_t *data, uint32_t available_bytes, std::string *value, int32_t min, int32_t max
+    const uint8_t *data, uint32_t available_bytes, std::string *value, size_t min, size_t max
 ) -> uint32_t;
-auto parse(uint8_t *const *data, uint32_t available_bytes, int32_t min, int32_t max)
+auto parse(uint8_t *const *data, uint32_t available_bytes, size_t min, size_t max)
     -> std::optional<std::string>;
 
 } // namespace cubic::protocol::primitives::string
