+++ libs/cubic-protocol/include/cubic-protocol/primitives/ushort.hpp
@@ -10,9 +10,10 @@ namespace cubic::protocol::primitives::ushort {
 auto parse(const uint8_t *data, uint32_t available_bytes, uint16_t *value) -> uint32_t;
 auto parse(uint8_t *const *data, uint32_t available_bytes, uint16_t *value)
     -> std::optional<uint16_t>;
-auto parse(const uint8_t *data, uint32_t available_bytes, uint16_t *value, int32_t min, int32_t max)
-    -> uint32_t;
-auto parse(uint8_t *const *data, uint32_t available_bytes, int32_t min, int32_t max)
+auto parse(
+    const uint8_t *data, uint32_t available_bytes, uint16_t *value, uint16_t min, uint16_t max
+) -> uint32_t;
+auto parse(uint8_t *const *data, uint32_t available_bytes, uint16_t min, uint16_t max)
     -> std::optional<uint16_t>;
 
 } // namespace cubic::protocol::primitives::ushort
